feat: Add a webpreference to disable pinch to zoom (#24)
OS-14345: Add a webpreference to disable pinch to zoom

In chromium pinch to zoom can be disabled globally using the command line switch
disable-pinch. BrightSign has a requirement to beable to disable pinch to zoom
at a browser window level. This patch adds enable_pinch_zoom as a webpreference
and sets its value in RenderWidgetHostViewEventHandler. The webpreference will
present on all platforms, but the change to set it is only present in aura (linux).
The webpreference will only be used if disable-pinch has not been set on the
command line.

Signed-off-by: Tariq Bashir <120014322+t-bashir-bs@users.noreply.github.com>