fix: Fix weston-keyboard backspace problem with existing text
Weston backend refuses to send deletion requests outside selection
area. Newly written text is in selection area, but old text is not.

This casues old text to be not deletable. Deleting outside the
selection area doesn't seem to be a problem as in EXO window manager
mentioned in the code comment.