feat: Add support for SetIgnoreMouseEvents on Wayland platform
Electron provides no implementation for SetIgnoreMouseEvents when using
Wayland. This patch provides uses the Chromium mouse lock functionality
to achieve a way of ignoring mouse events when using Wayland.
The change goes together with changes in our Electron App to listen to
the "focus" event and trigger a call to setIgnoreMouseEvents. This allows
us to workaround the issue of mouse lock being automatically disabled when
a BrowserWindow gains focus.
Using mouse lock for disabling mouse events works fine for our use case.