fix(nexus): Fix manual focus tracking with NexusWindow OS-18746
My previous attempt to do focus tracking on NexusWindow was based on
Mouse Grab/UnGrab. These events are not generated for touch events and
our Carson tests rely on those.

Replaced the focus tracking implementation to be similar to what we did
in Wayland. It now tracks all events which can activate a window and
sets the keyboard focus accordingly.