Fix reset when window.close is called OS-18755
If an html page calls window.close we end up getting a reset in
View::GetTransformRelativeTo. This issue has been found to be
caused by View::ConvertPointFromWidget being called after the
View has been freed (i.e. pointer use after being freed).
The exact sequence of events and what is the top level method
that results in calling ConvertPointFromWidget for a view after
its been freed is not possible to work out, due to the back
trace in all the dumps not showing where the call came from.
However, by adding traces in the code it was found that
WebContentsView::NonClientHitTest calls
View::ConvertPointFromWidget in this scenario and is always
done so in the error case after WebContentsView::WebContentsDestroyed.
Although this is not the root cause of the issue, this was a good
place to a fix where we check that api_web_contents_ is not null
in WebContentsView::NonClientHitTest before calling
View::ConvertPointFromWidget.