feat(text_tracks): Add support for in-band text tracks OS-18886
Added minimum interfaces available in WebMediaPlayerClient to inform
HtmlMediaElement with the text tracks found in the stream.

Majority of the logic for creating unique ids, and management of
tracks are carried over from QtWebEngine based Chromium. Unlike
the previous implementation, the ownership of the TextTrack
instances will be managed by HtmlMediaElement.